update tools & tool_resources for create-assistant & modify-assistant
